Don't install fonts on non-windowed servers

+++ bin/install_powerline_fonts.sh
@@ -1,12 +1,14 @@
 #!/bin/bash
 
-# Configure Directories
-powerline_dir="/tmp/powerline-fonts"
+# Only install fonts if in a Linux Windows Manager or MacOS
+if [[ (-z "$GDMSESSION") || (`uname` == "Darwin") ]]; then
+  powerline_dir="/tmp/powerline-fonts"
 
-# Download fonts
-git clone https://github.com/Lokaltog/powerline-fonts.git $powerline_dir
+  # Download fonts
+  git clone https://github.com/Lokaltog/powerline-fonts.git $powerline_dir
 
-# Install fonts
-$powerline_dir/install.sh
+  # Install fonts
+  $powerline_dir/install.sh
 
-rm -rf $powerline_dir
+  rm -rf $powerline_dir
+fi
